Run sshd-keygen on Fedora / CentOS 7 when host keys are missing
This *should* only run in Docker containers when sshd has never started. There's already a service that runs before sshd runs that does the generation, but that requires sshd to have already started once. That's not the case when we install into docker containers.

- Convert the flat library into a real helper under the Openssh::Helper namespace and send it to Resource
- Add new helpers to see if we're on a platform we keygen, return the command, and see if any of the hostkeys are missing. This avoids creating yet another attribute
