+++ recipes/default.rb
@@ -63,7 +63,6 @@ template '/etc/ssh/sshd_config' do
   owner 'root'
   group node['root_group']
   variables(options: openssh_server_options)
-  notifies :start, 'service[sshd-keygen]', :immediately
   notifies :run, 'execute[sshd-config-check]', :immediately
   notifies :restart, 'service[ssh]'
 end
@@ -73,10 +72,10 @@ execute 'sshd-config-check' do
   action :nothing
 end
 
-service 'sshd-keygen' do
-  supports [:restart, :reload, :status]
-  action :nothing
-  only_if { ::File.exist?('/usr/lib/systemd/system/sshd-keygen.service') }
+execute 'generate sshd host keys' do
+  command sshd_keygen_command
+  action :run
+  only_if { keygen_platform? && sshd_host_keys_missing? }
 end
 
 service 'ssh' do
